Putting arch file into a package.
Unfortunately, arch.zig and the kernel are dependant on each other. I
couldn't think of a way to have two pacakges directly depend on each other. I
made a file pluto.zig that collects and re-exports all the files in the
kernel. I made packages out of arch.zig and pluto.zig. kmain.zig then
imports those packges.

This works for just running the kernel but getting the tests to work
will be a bigger challenge.